requirements.txt
# Phase 0（知識庫建置）
pypdfium2>=4.30.0
Pillow>=10.0.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Phase 1（Bot MVP）— 暫時先裝也無妨
anthropic>=0.40.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0

# 測試
pytest>=8.0.0
